backend/requirements.txt
Flask==3.1.1
flask_cors==5.0.1
langgraph==0.4.7
matplotlib==3.10.3
moviepy==2.2.1
numpy==2.2.6
openai==1.82.1
openai_whisper==20240930
opencv_contrib_python==4.11.0.86
pyvista==0.45.2
spacy==3.8.4
